task can be deleted only if users created it

+++ src/AppBundle/Controller/TaskController.php
@@ -84,6 +84,8 @@ class TaskController extends Controller
      */
     public function deleteTaskAction(Task $task)
     {
+        $this->denyAccessUnlessGranted("edit", $task);
+
         $em = $this->getDoctrine()->getManager();
         $em->remove($task);
         $em->flush();
